Elasticsearch: Client refactor (#114745)
* split up client.go

* split up search_request.go

* remove double spaces
